Split off geometry & sat testing from math module
