+++ src/lib.rs
@@ -18,6 +18,7 @@ pub mod data_provider;
 // Workaround for https://github.com/rust-lang-nursery/error-chain/issues/254
 #[allow(deprecated)]
 pub mod errors;
+pub mod geometry;
 pub mod iterator;
 pub mod math;
 pub mod octree;
