+++ src/math.rs
@@ -13,13 +13,13 @@
 // limitations under the License.
 
 use cgmath::{
-    BaseFloat, BaseNum, Decomposed, Deg, EuclideanSpace, InnerSpace, Matrix3, Matrix4, Perspective,
-    Point3, Quaternion, Rotation, Transform, Vector3, Zero,
+    BaseFloat, BaseNum, Decomposed, Deg, EuclideanSpace, Matrix3, Point3, Quaternion, Rotation,
+    Vector3, Zero,
 };
-use collision::{Aabb, Aabb3, Contains, Relation};
+use collision::{Aabb3, Contains};
 use nav_types::{ECEF, WGS84};
 use num_traits::identities::One;
-use num_traits::Float;
+
 use s2::cell::Cell;
 use s2::cellid::CellID;
 use s2::cellunion::CellUnion;
@@ -28,6 +28,9 @@ use serde::{Deserialize, Serialize};
 use std::fmt;
 use std::ops::Mul;
 use std::str::FromStr;
+
+pub mod sat;
+pub use sat::*;
 
 /// Lower bound for distance from earth's center.
 /// See https://en.wikipedia.org/wiki/Earth_radius#Geophysical_extremes
@@ -195,95 +198,6 @@ where
         true
     }
 }
-#[derive(Debug, Clone)]
-pub struct Cube {
-    min: Point3<f64>,
-    edge_length: f64,
-}
-
-impl Cube {
-    pub fn bounding(aabb: &Aabb3<f64>) -> Self {
-        let edge_length = (aabb.max().x - aabb.min().x)
-            .max(aabb.max().y - aabb.min().y)
-            .max(aabb.max().z - aabb.min().z);
-        Cube {
-            min: aabb.min,
-            edge_length,
-        }
-    }
-
-    pub fn to_aabb3(&self) -> Aabb3<f64> {
-        Aabb3::new(self.min(), self.max())
-    }
-
-    pub fn new(min: Point3<f64>, edge_length: f64) -> Self {
-        Cube { min, edge_length }
-    }
-
-    pub fn edge_length(&self) -> f64 {
-        self.edge_length
-    }
-
-    pub fn min(&self) -> Point3<f64> {
-        self.min
-    }
-
-    pub fn max(&self) -> Point3<f64> {
-        Point3::new(
-            self.min.x + self.edge_length,
-            self.min.y + self.edge_length,
-            self.min.z + self.edge_length,
-        )
-    }
-
-    /// The center of the box.
-    pub fn center(&self) -> Vector3<f64> {
-        let min = self.min();
-        let max = self.max();
-        Vector3::new(
-            (min.x + max.x) / 2.,
-            (min.y + max.y) / 2.,
-            (min.z + max.z) / 2.,
-        )
-    }
-}
-
-// This guards against the separating axes being NaN, which may happen when the
-// orientation aligns with the unit axes.
-fn is_finite<S: BaseFloat>(vec: &Vector3<S>) -> bool {
-    vec.x.is_finite() && vec.y.is_finite() && vec.z.is_finite()
-}
-
-fn intersects_aabb3<S: BaseFloat>(
-    corners: &[Point3<S>],
-    separating_axes: &[Vector3<S>],
-    aabb: &Aabb3<S>,
-) -> bool {
-    // SAT algorithm
-    // https://gamedev.stackexchange.com/questions/44500/how-many-and-which-axes-to-use-for-3d-obb-collision-with-sat
-    for sep_axis in separating_axes.iter() {
-        // Project the cube and the box onto that axis
-        let mut cube_min_proj: S = Float::max_value();
-        let mut cube_max_proj: S = Float::min_value();
-        for corner in aabb.to_corners().iter() {
-            let corner_proj = corner.dot(*sep_axis);
-            cube_min_proj = cube_min_proj.min(corner_proj);
-            cube_max_proj = cube_max_proj.max(corner_proj);
-        }
-        // Project corners of the box onto that axis
-        let mut box_min_proj: S = Float::max_value();
-        let mut box_max_proj: S = Float::min_value();
-        for corner in corners.iter() {
-            let corner_proj = corner.dot(*sep_axis);
-            box_min_proj = box_min_proj.min(corner_proj);
-            box_max_proj = box_max_proj.max(corner_proj);
-        }
-        if box_min_proj > cube_max_proj || box_max_proj < cube_min_proj {
-            return false;
-        }
-    }
-    true
-}
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct Isometry3<S> {
@@ -377,193 +291,6 @@ impl<S: BaseFloat> Isometry3<S> {
             self.rotation.conjugate(),
             -(self.rotation.conjugate() * self.translation),
         )
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct Obb<S> {
-    query_from_obb: Isometry3<S>,
-    obb_from_query: Isometry3<S>,
-    half_extent: Vector3<S>,
-    corners: [Point3<S>; 8],
-    separating_axes: Vec<Vector3<S>>,
-}
-
-impl<S: BaseFloat> From<&Aabb3<S>> for Obb<S> {
-    fn from(aabb: &Aabb3<S>) -> Self {
-        Obb::new(
-            Isometry3::new(Quaternion::one(), EuclideanSpace::to_vec(aabb.center())),
-            aabb.dim() / (S::one() + S::one()),
-        )
-    }
-}
-
-impl<S: BaseFloat> From<Aabb3<S>> for Obb<S> {
-    fn from(aabb: Aabb3<S>) -> Self {
-        Self::from(&aabb)
-    }
-}
-
-impl<S: BaseFloat> Obb<S> {
-    pub fn new(query_from_obb: Isometry3<S>, half_extent: Vector3<S>) -> Self {
-        Obb {
-            obb_from_query: query_from_obb.inverse(),
-            half_extent,
-            corners: Obb::precompute_corners(&query_from_obb, &half_extent),
-            separating_axes: Obb::precompute_separating_axes(&query_from_obb.rotation),
-            query_from_obb,
-        }
-    }
-
-    pub fn transformed(&self, global_from_query: &Isometry3<S>) -> Self {
-        Self::new(global_from_query * &self.query_from_obb, self.half_extent)
-    }
-
-    fn precompute_corners(
-        query_from_obb: &Isometry3<S>,
-        half_extent: &Vector3<S>,
-    ) -> [Point3<S>; 8] {
-        let corner_from = |x, y, z| query_from_obb * &Point3::new(x, y, z);
-        [
-            corner_from(-half_extent.x, -half_extent.y, -half_extent.z),
-            corner_from(half_extent.x, -half_extent.y, -half_extent.z),
-            corner_from(-half_extent.x, half_extent.y, -half_extent.z),
-            corner_from(half_extent.x, half_extent.y, -half_extent.z),
-            corner_from(-half_extent.x, -half_extent.y, half_extent.z),
-            corner_from(half_extent.x, -half_extent.y, half_extent.z),
-            corner_from(-half_extent.x, half_extent.y, half_extent.z),
-            corner_from(half_extent.x, half_extent.y, half_extent.z),
-        ]
-    }
-
-    fn precompute_separating_axes(query_from_obb: &Quaternion<S>) -> Vec<Vector3<S>> {
-        let unit_x = Vector3::unit_x();
-        let unit_y = Vector3::unit_y();
-        let unit_z = Vector3::unit_z();
-        let rot_x = query_from_obb * unit_x;
-        let rot_y = query_from_obb * unit_y;
-        let rot_z = query_from_obb * unit_z;
-        let mut separating_axes = vec![unit_x, unit_y, unit_z];
-        for axis in &[
-            rot_x,
-            rot_y,
-            rot_z,
-            unit_x.cross(rot_x).normalize(),
-            unit_x.cross(rot_y).normalize(),
-            unit_x.cross(rot_z).normalize(),
-            unit_y.cross(rot_x).normalize(),
-            unit_y.cross(rot_y).normalize(),
-            unit_y.cross(rot_z).normalize(),
-            unit_z.cross(rot_x).normalize(),
-            unit_z.cross(rot_y).normalize(),
-            unit_z.cross(rot_z).normalize(),
-        ] {
-            let is_finite_and_non_parallel = is_finite(&axis)
-                && separating_axes.iter().all(|elem| {
-                    (elem - axis).magnitude() > S::default_epsilon()
-                        && (elem + axis).magnitude() > S::default_epsilon()
-                });
-            if is_finite_and_non_parallel {
-                separating_axes.push(*axis);
-            }
-        }
-        separating_axes
-    }
-}
-
-impl<S> PointCulling<S> for Obb<S>
-where
-    S: 'static + BaseFloat + Sync + Send,
-{
-    fn contains(&self, p: &Point3<S>) -> bool {
-        let Point3 { x, y, z } = &self.obb_from_query * p;
-        x.abs() <= self.half_extent.x
-            && y.abs() <= self.half_extent.y
-            && z.abs() <= self.half_extent.z
-    }
-    fn intersects_aabb3(&self, aabb: &Aabb3<S>) -> bool {
-        intersects_aabb3(&self.corners, &self.separating_axes, aabb)
-    }
-}
-
-impl<S> Cuboid<S> for Obb<S>
-where
-    S: BaseFloat,
-{
-    fn corners(&self) -> [Point3<S>; 8] {
-        self.corners
-    }
-}
-
-/// A frustum is defined in eye coordinates, where x points right, y points up,
-/// and z points against the viewing direction. This is not how e.g. OpenCV
-/// defines a camera coordinate system. To get from OpenCV camera coordinates
-/// to eye coordinates, you need to rotate 180 deg around the x axis before
-/// creating the perspective projection, see also the frustum unit test below.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct Frustum<S: BaseFloat> {
-    query_from_eye: Isometry3<S>,
-    clip_from_eye: Perspective<S>,
-    query_from_clip: Matrix4<S>,
-    frustum: collision::Frustum<S>,
-}
-
-impl<S: BaseFloat> Frustum<S> {
-    pub fn new(query_from_eye: Isometry3<S>, clip_from_eye: Perspective<S>) -> Self {
-        let eye_from_query: Decomposed<Vector3<S>, Quaternion<S>> = query_from_eye.inverse().into();
-        let clip_from_query =
-            Matrix4::<S>::from(clip_from_eye) * Matrix4::<S>::from(eye_from_query);
-        let query_from_clip = clip_from_query.inverse_transform().unwrap();
-        let frustum = collision::Frustum::from_matrix4(clip_from_query).unwrap();
-        Frustum {
-            query_from_eye,
-            clip_from_eye,
-            query_from_clip,
-            frustum,
-        }
-    }
-
-    pub fn transformed(&self, global_from_query: &Isometry3<S>) -> Self {
-        Self::new(global_from_query * &self.query_from_eye, self.clip_from_eye)
-    }
-}
-
-impl<S> PointCulling<S> for Frustum<S>
-where
-    S: 'static + BaseFloat + Sync + Send,
-{
-    fn contains(&self, point: &Point3<S>) -> bool {
-        match self.frustum.contains(point) {
-            Relation::Cross => true,
-            Relation::In => true,
-            Relation::Out => false,
-        }
-    }
-    fn intersects_aabb3(&self, aabb: &Aabb3<S>) -> bool {
-        match self.frustum.contains(aabb) {
-            Relation::Cross => true,
-            Relation::In => true,
-            Relation::Out => false,
-        }
-    }
-}
-
-impl<S> Cuboid<S> for Frustum<S>
-where
-    S: BaseFloat,
-{
-    fn corners(&self) -> [Point3<S>; 8] {
-        let corner_from = |x, y, z| self.query_from_clip.transform_point(Point3::new(x, y, z));
-        [
-            corner_from(-S::one(), -S::one(), -S::one()),
-            corner_from(-S::one(), -S::one(), S::one()),
-            corner_from(-S::one(), S::one(), -S::one()),
-            corner_from(-S::one(), S::one(), S::one()),
-            corner_from(S::one(), -S::one(), -S::one()),
-            corner_from(S::one(), -S::one(), S::one()),
-            corner_from(S::one(), S::one(), -S::one()),
-            corner_from(S::one(), S::one(), S::one()),
-        ]
     }
 }
